backend/requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.3

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0

# AWS S3
boto3==1.34.10

# AI Services
anthropic==0.18.1
openai==1.12.0

# GitHub
PyGithub==2.1.1
GitPython==3.1.41

# Utilities
requests==2.31.0
python-magic==0.4.27
uuid==1.30

# Document Export
python-docx==1.1.0
markdown2==2.4.12
# weasyprint==60.2  # Optional - requires GTK libraries (difficult on Windows)
# For PDF export on Windows, see: https://github.com/tschoonj/GTK-for-Windows-Runtime-Environment-Installer/releases

# Background Tasks (optional)
celery==5.3.4
redis==5.0.1

# Development
gunicorn==21.2.0

